requirements.txt
fastapi-0.109.2
uvicorn[standard]
shortuuid